3.6

+++ aeserver/src/main/java/nl/team12/amsterdamevents/aeserver/app/rest/AEventsController.java
@@ -1,6 +1,8 @@
 package nl.team12.amsterdamevents.aeserver.app.rest;
 
 import com.fasterxml.jackson.annotation.JsonView;
+import nl.team12.amsterdamevents.aeserver.app.exceptions.PreConditionFailed;
+import nl.team12.amsterdamevents.aeserver.app.exceptions.ResourceNotFoundException;
 import nl.team12.amsterdamevents.aeserver.app.models.AEvent;
 import nl.team12.amsterdamevents.aeserver.app.repositories.AEventsRepository;
 import nl.team12.amsterdamevents.aeserver.app.views.AEventView;
@@ -8,11 +10,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @CrossOrigin(origins = "http://localhost:8080")
@@ -26,10 +28,11 @@ public class AEventsController {
         this.aEventsRepository = aEventsRepository;
     }
 
-    private URI getLocationURI(Long id) {
-        return ServletUriComponentsBuilder.
-                fromCurrentRequest().path("/{id}").
-                buildAndExpand(id).toUri();
+    public URI getLocationURI(Long id) {
+        return ServletUriComponentsBuilder.fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(id)
+                .toUri();
     }
 
     // GET mapping which gets all the aEvents
@@ -38,26 +41,29 @@ public class AEventsController {
         return aEventsRepository.findAll();
     }
 
+    // GET mapping of the summary of the aEvents
     @GetMapping("/aevents/summary")
     @JsonView(value = AEventView.SummaryView.class)
-    public List<AEvent> getAEventsSummary(){
+    public List<AEvent> getAEventsSummary() {
         return aEventsRepository.findAll();
     }
 
     // GET mapping which uses the AEvent id to deliver the aEvent that is identified by the specified path variable
+    // Exception if request consists of a non-existing id
     @GetMapping("/aevents/{id}")
-    public AEvent getAEventById(@PathVariable Long id) throws Exception {
+    public AEvent getAEventById(@PathVariable Long id) {
         AEvent aEvent = aEventsRepository.findById(id);
         if (aEvent == null) {
-            throw new Exception("AEvent id= " + id + " not found.");
-        } else {
-            return aEvent;
+            throw new ResourceNotFoundException("AEvent id=" + id + " not found");
         }
+        return aEvent;
     }
 
     // POST mapping which adds a new aEvent to the repository
+    // Exception if id is different from the id that is provided with the aEvent in the request body
+    // works without @RequestBody?
     @PostMapping("/aevents")
-    public ResponseEntity<AEvent> saveAEvent(@RequestBody AEvent aEvent) {
+    public ResponseEntity<AEvent> saveAEvent(AEvent aEvent) {
         AEvent savedAEvent = aEventsRepository.save(aEvent);
 
         URI location = getLocationURI(savedAEvent.getId());
@@ -65,20 +71,27 @@ public class AEventsController {
     }
 
     // PUT mapping which updates/replaces the stored aEvent identified by id
+    // Exception if id is different from the id that is provided with the aEvent in the request body
+    // works without @RequestBody?
     @PutMapping("/aevents/{id}")
-    public ResponseEntity<AEvent> updateAEvent(@PathVariable Long id, @RequestBody AEvent aEvent) {
-        if (id.equals(aEvent.getId())) {
+    public ResponseEntity<AEvent> updateAEvent(@PathVariable Long id, AEvent aEvent) {
+        if (Objects.equals(id, aEvent.getId())) {
             AEvent savedAEvent = aEventsRepository.save(aEvent);
             return ResponseEntity.accepted().body(savedAEvent);
         }
-        throw new ResponseStatusException(HttpStatus.PRECONDITION_FAILED, "AEvent not found with id: " + id);
+        throw new PreConditionFailed("AEvent-id=" + aEvent.getId() + " does not match path parameter=" + id);
     }
 
     // DELETE mapping which uses the id to remove the identified aEvent from the repository
+    // Exception if request consists of a non-existing id
     @DeleteMapping("/aevents/{id}")
     public ResponseEntity<URI> deleteAEventById(@PathVariable Long id) {
-        aEventsRepository.deleteById(id);
-        return ResponseEntity.ok(ServletUriComponentsBuilder.fromCurrentRequest().path("").build().toUri());
+        if (!aEventsRepository.deleteById(id)) {
+            throw new ResourceNotFoundException("AEvent id=" + id + " not found");
+        } else {
+            aEventsRepository.deleteById(id);
+            return new ResponseEntity(id, HttpStatus.NO_CONTENT);
+        }
     }
 }
